Working on switching over to a CMake-based build and adding Windows support (#98)
- Switched to CMake-based config and removed legacy build system.
- Added Windows support.
- Tracking latest bls-signatures version.
- Updated README with new build instructions.
- Updated LICESNE copyrights.

+++ src/ChiaConsts.h
@@ -138,6 +138,6 @@ inline size_t CalculateParkSize( TableId tableId )
 // values of k, we need extra space to account for the additional variability.
 constexpr inline static size_t CalculateC3Size()
 {
-    return CDiv( kC3BitsPerEntry * kCheckpoint1Interval, 8 );
+    return (size_t)CDiv( kC3BitsPerEntry * kCheckpoint1Interval, 8 );
 }
